+++ CHANGELOG.md
@@ -46,6 +46,7 @@ Section Order:
 ### Added
 
 - New Cogs
+  - `Routes` cog to find your way through New Eden
   - `WhereIs` cog to find where you lost your stuff
 
 ## [3.0.0] - 2026-06-07