+++ CHANGELOG.md
@@ -43,6 +43,8 @@ Section Order:
 
 <!-- Your changes go here -->
 
+## [3.1.0] - 2026-06-16
+
 ### Added
 
 - New Cogs
@@ -394,6 +396,7 @@ Section Order:
 [2.1.1]: https://github.com/ppfeufer/tn-nt-discordbot-cogs/compare/v2.1.0...v2.1.1 "v2.1.1"
 [2.2.0]: https://github.com/ppfeufer/tn-nt-discordbot-cogs/compare/v2.1.1...v2.2.0 "v2.2.0"
 [3.0.0]: https://github.com/ppfeufer/tn-nt-discordbot-cogs/compare/v2.2.0...v3.0.0 "v3.0.0"
-[in development]: https://github.com/ppfeufer/tn-nt-discordbot-cogs/compare/v3.0.0...HEAD "In Development"
+[3.1.0]: https://github.com/ppfeufer/tn-nt-discordbot-cogs/compare/v3.0.0...v3.1.0 "v3.1.0"
+[in development]: https://github.com/ppfeufer/tn-nt-discordbot-cogs/compare/v3.1.0...HEAD "In Development"
 [keep a changelog]: http://keepachangelog.com/ "Keep a Changelog"
 [semantic versioning]: http://semver.org/ "Semantic Versioning"